[NONPR-4616][DC] moving error case test into seperate file to allow some seperation between resources to allow the negative case to be true
