+++ src/it/scala/uk/gov/hmrc/nonrep/attachment/ServiceIntSpec.scala
@@ -5,7 +5,7 @@ import org.apache.pekko.actor.typed.ActorSystem
 import org.apache.pekko.actor.typed.scaladsl.adapter.*
 import org.apache.pekko.http.scaladsl.Http
 import org.apache.pekko.http.scaladsl.model.{HttpRequest, HttpResponse, StatusCodes}
-import org.scalatest.Inside
+import org.scalatest.{BeforeAndAfterEach, Inside}
 import org.scalatest.time.{Millis, Seconds, Span}
 import uk.gov.hmrc.nonrep.BuildInfo
 import uk.gov.hmrc.nonrep.attachment.server.{NonrepMicroservice, ServiceConfig}
@@ -53,19 +53,6 @@ class ServiceIntSpec extends BaseSpec with Inside {
         res.status shouldBe StatusCodes.OK
         whenReady(entityToString(res.entity)) { body =>
           body shouldBe "pong"
-        }
-      }
-    }
-
-    "return a 500 response for GET requests to service /ping endpoint when attachments processor has stopped" in {
-      import scala.jdk.FutureConverters.*
-      server.attachmentsProcessor.asJava.toCompletableFuture.cancel(true)
-      Thread.sleep(1000)
-      val responseFuture: Future[HttpResponse] = Http(system).singleRequest(HttpRequest(uri = s"$hostUrl/${config.appName}/ping"))
-      whenReady(responseFuture) { res =>
-        res.status shouldBe StatusCodes.InternalServerError
-        whenReady(entityToString(res.entity)) { body =>
-          body shouldBe "Processing of attachments is finished"
         }
       }
     }
